# swagger.yaml
host: localhost:25100
info:
  contact: {}
  description:
    Digital Publishing resumable file upload service that handles writing
    to S3. It updates images through the CMS.
  license:
    name: Open Government Licence v3.0
    url: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
  title: dp-upload-service
  version: 1.0.0
paths:
  /health:
    get:
      consumes:
        - application/json
      description: returns a health check
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload
  /upload:
    get:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
        - multipart/form-data
      description: checks to see if a chunk has been uploaded
      parameters:
        - in: formData
          name: aliasName
          description: An alias name for the file
          required: true
          type: string
        - in: formData
          name: chunkNumber
          description: The index of the current chunk being uploaded, the first chunk is 1 not base 0
          required: true
          type: integer
        - in: formData
          name: chunkSize
          description: The size of each chunk in bytes
          required: true
          type: integer
        - in: formData
          name: currentChunkSize
          description: The size of a current chunk
          required: true
          type: integer
        - in: formData
          name: fileName
          description: The name of the file being uploaded
          required: true
          type: string
        - in: formData
          name: identifier
          description: A unique identifier for the file upload session
          required: true
          type: string
        - in: formData
          name: relativePath
          description: The path where the file will be stored
          required: true
          type: string
        - in: formData
          name: totalChunks
          description: The total number of chunks the file is divided into
          required: true
          type: integer
        - in: formData
          name: totalSize
          description: The total size of the file in bytes
          required: true
          type: integer
        - in: formData
          name: type
          description: The MIME type of the file being uploaded
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload
    post:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
        - multipart/form-data
      description: handles the uploading of a file to AWS S3
      parameters:
        - in: formData
          name: aliasName
          description: An alias name for the file
          required: true
          type: string
        - in: formData
          name: chunkNumber
          description: The index of the current chunk being uploaded, the first chunk is 1 not base 0
          required: true
          type: integer
        - in: formData
          name: chunkSize
          description: The size of each chunk in bytes
          required: true
          type: integer
        - in: formData
          name: currentChunkSize
          description: The size of a current chunk
          required: true
          type: integer
        - in: formData
          name: fileName
          description: The name of the file being uploaded
          required: true
          type: string
        - in: formData
          name: identifier
          description: A unique identifier for the file upload session
          required: true
          type: string
        - in: formData
          name: relativePath
          description: The path where the file will be stored
          required: true
          type: string
        - in: formData
          name: totalChunks
          description: The total number of chunks the file is divided into
          required: true
          type: integer
        - in: formData
          name: totalSize
          description: The total size of the file in bytes
          required: true
          type: integer
        - in: formData
          name: type
          description: The MIME type of the file being uploaded
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload
  /upload/{id}:
    get:
      consumes:
        - application/json
      description:
        returns an S3 URL for a requested path, and the client's region
        and bucket name
      parameters:
        - description: S3 object key
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload
  /upload-new:
    post:
      consumes:
        - multipart/form-data
      description: Handles the uploading of a file in chunks to AWS S3
      parameters:
        - in: formData
          name: aliasName
          description: An alias name for the file
          required: true
          type: string
        - in: formData
          name: collectionId
          description: The ID of the collection to which the file belongs
          required: true
          type: string
        - in: formData
          name: bundleId
          description: The ID of the bundle to which the file belongs
          required: false
          type: string
        - in: formData
          name: file
          description: Specifies the file to be uploaded
          required: true
          type: file
        - in: formData
          name: isPublishable
          description: A boolean indicating whether the file is publishable
          required: true
          type: string
        - in: formData
          name: licence
          description: The type of license associated with the file
          required: true
          type: string
        - in: formData
          name: licenceUrl
          description: A URL linking to the license associated with the file
          required: true
          type: string
        - in: formData
          description: The path where the file will be stored
          name: path
          required: true
          type: string
        - in: formData
          name: resumableChunkNumber
          description: The index of the current chunk being uploaded, the first chunk is 1 not base 0 
          required: true
          type: integer
        - in: formData
          name: resumableChunkSize
          description: The size of each chunk in bytes
          required: true
          type: integer
        - in: formData
          name: resumableFilename
          description: The name of the file being uploaded
          required: true
          type: string
        - in: formData
          name: resumableIdentifier
          description: A unique identifier for the file upload session
          required: true
          type: string
        - in: formData
          name: resumableType
          description: The MIME type of the file being uploaded
          required: true
          type: string
        - in: formData
          name: resumableTotalChunks
          description: The total number of chunks the file is divided into
          required: true
          type: integer
        - in: formData
          name: resumableTotalSize
          description: The total size of the file in bytes
          required: true
          type: integer
        - in: formData
          name: datasetId
          description: The dataset ID that the file relates to
          required: true
          type: string
        - in: formData
          name: edition
          description: The edition that the file relates to
          required: true
          type: string
        - in: formData
          name: version
          description: The version that the file relates to
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload-new
  /upload-new/files/{path}/{file-name}/status:
    get:
      consumes:
        - application/json
      description: Gets the status of a file upload
      parameters:
        - in: path
          name: path
          description: The path where the file will be stored
          required: true
          type: string
        - in: path
          name: file-name
          description: The name of the file being uploaded
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              metadata:
                type: object
                properties:
                  path:
                    type: string
                  is_publishable:
                    type: boolean
                  collection_id:
                    type: string
                  title:
                    type: string
                  size_in_bytes:
                    type: integer
                  type:
                    type: string
                  licence:
                    type: string
                  licence_url:
                    type: string
                  state:
                    type: string
                  etag:
                    type: string
              file_content:
                type: object
                properties:
                  valid:
                    type: boolean
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - upload-new
schemes:
  - http
swagger: "2.0"